requirements.txt
PyPDF2
python-docx
nltk
scikit-learn
numpy